fixed style issue

+++ building_sim_plugins/building_plugins_common/include/building_sim_common/slotcar_common.hpp
@@ -280,7 +280,7 @@ private:
 
   double compute_charge(const double run_time) const;
 
-  int get_next_noncolinear_target(const rclcpp::Time current_time);
+  size_t get_next_noncolinear_target(const rclcpp::Time current_time);
 
   double compute_discharge(
     const Eigen::Vector3d lin_vel, const double ang_vel,
